Implemented a new configuration module
I change the config.py so it is now possible set defaults values inside
every cli/cmd/, and I added code to load a config file.
The file needs to be in the INI forma, and I added code to load a config
file.
The file needs to be in the INI format..