Now doto can add appointments
The apppointment class waas added and everything to store an appointment
inside the db.

In addition a new cli command was added with which now tasks can be
added. Maybe this command will be joined with the current add command to
something like

   doto add task ...
or
   doto add apmt ...

But that is future me's problem.